rename fine tuning apis
